+++ lib/VitableConnect/types/organization_user_role.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module VitableConnect
+  module Types
+    module OrganizationUserRole
+      extend VitableConnect::Internal::Types::Enum
+
+      ADMIN = "ADMIN"
+      OPERATIONS = "OPERATIONS"
+      SALES = "SALES"
+      ENROLLMENT_AGENT = "ENROLLMENT_AGENT"
+    end
+  end
+end